feat(clapi): add possibility to grant access to children menu (or not)

+++ www/class/centreon-clapi/centreonACLMenu.class.php
@@ -55,7 +55,6 @@ class CentreonACLMenu extends CentreonObject
     const LEVEL_2 = 1;
     const LEVEL_3 = 2;
     const LEVEL_4 = 3;
-    const PARENT_MENU_NOT_DEFINED = "Parent menu must be defined";
     protected $relObject;
     protected $aclGroupObj;
     protected $topologyObj;
@@ -163,25 +162,26 @@ class CentreonACLMenu extends CentreonObject
     protected function splitParams($parameters)
     {
         $params = explode($this->delim, $parameters);
-        if (count($params) < 2) {
+        if (count($params) < 3) {
             throw new CentreonClapiException(self::MISSINGPARAMETER);
         }
         $aclMenuId = $this->object->getIdByParameter($this->object->getUniqueLabelField(), array($params[0]));
         if (!count($aclMenuId)) {
             throw new CentreonClapiException(self::OBJECT_NOT_FOUND . ":" . $params[0]);
         }
+        $processChildren = ($params[1] == '0') ? false : true;
         $levels = array();
         $menus = array();
-        $toopologies = array();
-        $levels[self::LEVEL_1] = $params[1];
+        $topologies = array();
+        $levels[self::LEVEL_1] = $params[2];
         if (isset($params[2])) {
-            $levels[self::LEVEL_2] = $params[2];
+            $levels[self::LEVEL_2] = $params[3];
         }
         if (isset($params[3])) {
-            $levels[self::LEVEL_3] = $params[3];
+            $levels[self::LEVEL_3] = $params[4];
         }
         if (isset($params[4])) {
-            $levels[self::LEVEL_4] = $params[4];
+            $levels[self::LEVEL_4] = $params[5];
         }
         foreach ($levels as $level => $menu) {
             if ($menu) {
@@ -235,7 +235,7 @@ class CentreonACLMenu extends CentreonObject
                 break;
             }
         }
-        return array($aclMenuId[0], $menus, $topologies);
+        return array($aclMenuId[0], $menus, $topologies, $processChildren);
     }
 
     /**
@@ -273,8 +273,11 @@ class CentreonACLMenu extends CentreonObject
      * @param int $parentTopologyId
      * @return void
      */
-    protected function processChildrenOf($action = "grant", $aclMenuId = null, $parentTopologyId = null)
-    {
+    protected function processChildrenOf(
+        $action = "grant",
+        $aclMenuId = null,
+        $parentTopologyId = null
+    ) {
         $sql = "SELECT topology_id, topology_page FROM topology WHERE topology_parent = ?";
         $res = $this->db->query($sql, array($parentTopologyId));
         $rows = $res->fetchAll();
@@ -318,7 +321,7 @@ class CentreonACLMenu extends CentreonObject
      */
     public function grantRw($parameters)
     {
-        list($aclMenuId, $menus, $topologies) = $this->splitParams($parameters);
+        list($aclMenuId, $menus, $topologies, $processChildren) = $this->splitParams($parameters);
         foreach ($menus as $level => $menuId) {
             $this->db->query(
                 "DELETE FROM acl_topology_relations WHERE acl_topo_id = ? AND topology_topology_id = ?",
@@ -328,7 +331,7 @@ class CentreonACLMenu extends CentreonObject
                 "INSERT INTO acl_topology_relations (acl_topo_id, topology_topology_id) VALUES (?, ?)",
                 array($aclMenuId, $menuId)
             );
-            if (!isset($menus[$level + 1]) && $level != self::LEVEL_4) {
+            if ($processChildren && !isset($menus[$level + 1]) && $level != self::LEVEL_4) {
                 $this->processChildrenOf("grant", $aclMenuId, $topologies[$level]);
             }
         }
@@ -343,7 +346,7 @@ class CentreonACLMenu extends CentreonObject
      */
     public function grantRo($parameters)
     {
-        list($aclMenuId, $menus, $topologies) = $this->splitParams($parameters);
+        list($aclMenuId, $menus, $topologies, $processChildren) = $this->splitParams($parameters);
         foreach ($menus as $level => $menuId) {
             $this->db->query(
                 "DELETE FROM acl_topology_relations WHERE acl_topo_id = ? AND topology_topology_id = ?",
@@ -353,7 +356,7 @@ class CentreonACLMenu extends CentreonObject
                 "INSERT INTO acl_topology_relations (acl_topo_id, topology_topology_id, access_right) VALUES (?, ?, 2)",
                 array($aclMenuId, $menuId)
             );
-            if (!isset($menus[$level + 1]) && $level != self::LEVEL_4) {
+            if ($processChildren && !isset($menus[$level + 1]) && $level != self::LEVEL_4) {
                 $this->processChildrenOf("grantro", $aclMenuId, $topologies[$level]);
             }
         }
@@ -368,9 +371,9 @@ class CentreonACLMenu extends CentreonObject
      */
     public function revoke($parameters)
     {
-        list($aclMenuId, $menus, $topologies) = $this->splitParams($parameters);
+        list($aclMenuId, $menus, $topologies, $processChildren) = $this->splitParams($parameters);
         foreach ($menus as $level => $menuId) {
-            if (!isset($menus[$level + 1])) {
+            if ($processChildren && !isset($menus[$level + 1])) {
                 $this->db->query(
                     "DELETE FROM acl_topology_relations WHERE acl_topo_id = ? AND topology_topology_id = ?",
                     array($aclMenuId, $menuId)
@@ -420,8 +423,10 @@ class CentreonACLMenu extends CentreonObject
 
         $grantedMenu = '';
 
-        $grantedMenuTpl = $this->action . $this->delim . '%s' . $this->delim .
+        $grantedMenuTpl = $this->action . $this->delim .
+            '%s' . $this->delim .
             $aclTopoName . $this->delim .
+            '%s' . $this->delim .
             '%s' . $this->delim . "\n";
 
         $grantedPossibilities = array(
@@ -432,7 +437,7 @@ class CentreonACLMenu extends CentreonObject
         $queryAclMenuRelations = 'SELECT t.topology_page, t.topology_id, t.topology_name, atr.access_right ' .
             'FROM acl_topology_relations atr, topology t ' .
             'WHERE atr.topology_topology_id = t.topology_id ' .
-            "AND atr.access_right <> '0'" .
+            "AND atr.access_right <> '0' " .
             'AND atr.acl_topo_id = ?';
 
         $grantedTopologyList = $this->db->fetchAll($queryAclMenuRelations, array($aclTopoId));
@@ -446,6 +451,7 @@ class CentreonACLMenu extends CentreonObject
             $grantedMenu .= sprintf(
                 $grantedMenuTpl,
                 $grantedPossibilities[$grantedTopology['access_right']],
+                '0',
                 $grantedTopologyBreadCrumb
             );
         }
